# server/internal/docs/swagger.yaml
basePath: /
definitions:
...
  github_com_hybrid-app_server_internal_httpx.Envelope:
    properties:
      code:
        description: 0 = 成功，非 0 = 业务错误码（沿用 HTTP 状态码语义）
        type: integer
      data:
        description: 业务数据
      message:
        description: 人类可读信息
        type: string
    type: object
...
  github_com_hybrid-app_server_internal_service.PushCampaignInput:
    properties:
      body:
        type: string
      deeplinkPath:
        type: string
      extraData:
        additionalProperties:
          type: string
        type: object
      imageUrl:
        type: string
      name:
        type: string
      targetAppIds:
        items:
          type: string
        type: array
      title:
        type: string
    type: object
...
  internal_handler.registerTokenReq:
    properties:
      appId:
        type: string
      model:
        type: string
      palcode:
        type: string
      platform:
        type: string
      token:
        type: string
    type: object
...
  internal_handler.scheduleCampaignReq:
    properties:
      scheduledAt:
        description: ISO8601
        type: string
    type: object
...
  internal_handler.sendCampaignReq:
    properties:
      dryRun:
        type: boolean
    type: object
...
info:
  contact: {}
  description: hybrid-app 多渠道打包后台：渠道 CRUD / 图标管线 / 域名配置 / 运行时下发。
  title: 渠道中台 API
  version: "1.0"
paths:
...
  /api/app/push/register-token:
    post:
      consumes:
      - application/json
      parameters:
      - description: token 注册请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.registerTokenReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_hybrid-app_server_internal_httpx.Envelope'
      summary: APK 上报 FCM 设备 token（公开端点）
      tags:
      - push
...
  /api/push/audience:
    get:
      parameters:
      - description: 逗号分隔的 applicationId 列表
        in: query
        name: appIds
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_hybrid-app_server_internal_httpx.Envelope'
      security:
      - BearerAuth: []
      summary: 预估目标活跃设备数（viewer）
      tags:
      - push
  /api/push/campaigns:
    get:
      parameters:
      - description: 按品牌 code 筛选
        in: query
        name: brand
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_hybrid-app_server_internal_httpx.Envelope'
      security:
      - BearerAuth: []
      summary: 推送活动列表（viewer）
      tags:
      - push
    post:
      consumes:
      - application/json
      parameters:
      - description: 活动内容
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_hybrid-app_server_internal_service.PushCampaignInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_hybrid-app_server_internal_httpx.Envelope'
      security:
      - BearerAuth: []
      summary: 创建推送活动草稿（operator）
      tags:
      - push
  /api/push/campaigns/{id}:
    get:
      parameters:
      - description: 活动 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_hybrid-app_server_internal_httpx.Envelope'
      security:
      - BearerAuth: []
      summary: 推送活动详情（viewer）
      tags:
      - push
    put:
      consumes:
      - application/json
      parameters:
      - description: 活动 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新内容
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_hybrid-app_server_internal_service.PushCampaignInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_hybrid-app_server_internal_httpx.Envelope'
      security:
      - BearerAuth: []
      summary: 修改推送活动草稿（operator，仅 draft 可改）
      tags:
      - push
  /api/push/campaigns/{id}/schedule:
    post:
      consumes:
      - application/json
      parameters:
      - description: 活动 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 定时时间 ISO8601
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.scheduleCampaignReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_hybrid-app_server_internal_httpx.Envelope'
      security:
      - BearerAuth: []
      summary: 设置推送活动定时发送（operator）
      tags:
      - push
  /api/push/campaigns/{id}/send:
    post:
      consumes:
      - application/json
      parameters:
      - description: 活动 ID
        in: path
        name: id
        required: true
        type: integer
      - description: dryRun 标志
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.sendCampaignReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_hybrid-app_server_internal_httpx.Envelope'
      security:
      - BearerAuth: []
      summary: 立即发送推送活动（operator）；dryRun=true 走完统计但不真发
      tags:
      - push
...
  /api/push/status:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_hybrid-app_server_internal_httpx.Envelope'
      security:
      - BearerAuth: []
      summary: FCM 配置状态（viewer）
      tags:
      - push
  /api/push/upload-image:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 图片文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_hybrid-app_server_internal_httpx.Envelope'
      security:
      - BearerAuth: []
      summary: 上传推送图片（operator）
      tags:
      - push
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"